pyproject.toml
[tool.poetry]
name = "studio-caju"
version = "0.1.0"
description = ""
authors = ["Victor Contreras <contreras3991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.39"}
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
alembic = "^1.14.1"
fastapi = {extras = ["standard"], version = "^0.115.8"}
pre-commit = "^4.1.0"
pyjwt = "^2.10.1"
tzdata = "^2025.1"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
loguru = "^0.7.3"
psycopg = {extras = ["binary"], version = "^3.2.5"}
mkdocs = "^1.6.1"
aiosqlite = "^0.21.0"
asyncpg = "^0.30.0"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
ruff = "^0.9.6"
pytest-asyncio = "^0.25.3"
freezegun = "^1.5.1"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev src/api/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html'
docs-serve = "mkdocs serve -f mkdocs.yml"
docs-build = "mkdocs build -f mkdocs.yml"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
